Move database to common module
