+++ common/src/main/java/io/homeassistant/companion/android/database/authentication/AuthenticationDao.kt
