+++ common/src/main/java/io/homeassistant/companion/android/database/notification/NotificationItem.kt
