+++ common/src/main/java/io/homeassistant/companion/android/database/sensor/SensorDao.kt
