+++ common/src/main/java/io/homeassistant/companion/android/database/sensor/SensorWithSettings.kt
