+++ common/src/main/java/io/homeassistant/companion/android/database/widget/CameraWidgetDao.kt
