+++ common/src/main/res/drawable-hdpi/ic_stat_ic_notification_blue.png
