+++ common/src/main/res/drawable-mdpi/ic_stat_ic_notification_blue.png
