+++ common/src/main/res/drawable-xhdpi/ic_stat_ic_notification_blue.png
