+++ common/src/main/res/drawable-xxhdpi/ic_stat_ic_notification.png
