+++ common/src/main/res/drawable-xxxhdpi/ic_stat_ic_notification_blue.png
